+++ CHANGELOG.rst
@@ -6,7 +6,7 @@ Changelog
 
 * Add new method :meth:`pyvcf.VcfFrame.add_af`.
 * Add new method :meth:`pyvcf.VcfFrame.extract`.
-* Deprecate methods :meth:`pyvep.filter_af/biotype`.
+* Deprecate methods :meth:`pyvep.filter_af/biotype/nothas/impact`.
 * Add new method :meth:`pyvep.filter_query`.
 * :issue:`19`: Add new command :command:`vep_query`.